pull out route logic into utils, move create grant button
